Add files via upload
Program prompts users to enter a beginning population size that must be greater than or equal to 9 and an ending population size that must be bigger than the beginning  population size. Program then calculates the amount of years required to get from beginning population size to ending population size assuming that for n species, n/3 are born and n/4 pass away.